feat(UVa-696): final solution - PASSED
Used the geometric properties of the board to compute the solution by
differentiating between the 3 cases:

- Case 1: n = 1 or m = 1
  Since to move or attack a knight needs at least one row or column
different than its position for displacement, we can place knights on
all the available cells, they cannot attack each other.

- Case 2: n == 2
  The same pattern is repeated by column or row of 4 in the maximum
dimension (m), and the number of knights to put on the remainder rows or
columns is deduced.

- Case 3: n > 2 (General case)
  Since a knight moves from one color to the oposite, we can place a
knight on the maximum cells of the same color, they will not be able to
attack each other if they are on the same cell color. An integer trick
(+1) is used to find the correct maximum cell color where in cases such
as N=5 and M=5 we have 25 cells and dividing by would give the answer
12, but we can have a maximum of 13 cells of the same color on a 5x5
board to place knights. That is the reason of the (+1) in the result
(n*m + 1) / 2 for the number of knights

Signed-off-by: Neil Taison Rigaud <ntsrigaud14@gmail.com>